requirements.txt
asgiref==3.4.1
Django==3.1.3
djangorestframework==3.12.2
mysqlclient==2.0.1
django-mysql==3.9
django-cors-headers==3.5.0
pytz==2021.3
sqlparse==0.4.2